[BEAM-3147][NexmarkLauncher] Extract NexmarkPerfAnalyzer out of NexmarkLauncher
Refactor to make NexmarkLauncher smaller
